requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
ecdsa==0.15
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.3
msal==1.1.0
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.11.0
python-jose==3.1.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
